detect correctly whether a file system is currently mounted (bsc#1105227)
This patch (a) ensures that the mount point is in a 'canonical' form.

And (b) the code keeps working if the mount point in /etc/fstab is incorrect.

The code used to prefer the mount point from /etc/fstab even if the file
system was actually mounted elsewhere. Now it uses the value from /proc/mounts
and logs a warning if /etc/fstab disagrees.