LUKS2: add AEAD integrity option (PM-3419)
LUKS2 support "integrity" and "data authentication" since 2.0 in
experimental mode.  This is required when we want to detect changes out
of band in the devices.

Recently the kernel is supporting good AEAD family of algorithms that
will guarantee unauthorized changes in the device (instead of decrypting
random data) for big disks.

For integrity it is recommended "aegis128-random" with AEAD, as other
family (like chacha20-random with poly1305, or aes-xts-plain64 with
hmac-sha256) have performance problems, or are not supporting large
enought devices.  A check is implemented to validate the data.

This patch add a new element, "integrity", that for now can only have
"aead" value.  When other methods are added, the check can be changed
without changing the XML data, nor the parser.  If integrity is set,
also it is checked that "aegis128-random" cipher algorithm is used.

Signed-off-by: Alberto Planas <aplanas@suse.com>